some gcc fixes

+++ src/framework/math/fbm.cpp
@@ -37,7 +37,7 @@ float fbm::hash(const vec2& p)
 
 float fbm::noise2(const vec2& st)
 {
-   vec2 i{floor(st.x), floor(st.y)};
+   vec2 i{static_cast<float>(floor(st.x)), static_cast<float>(floor(st.y))};
    vec2 f{fract(st.x), fract(st.y)};
 
    // four corners in 2D of a tile
@@ -58,7 +58,7 @@ float fbm::random1(const vec2& st)
 
 float fbm::noise1(const vec2& st)
 {
-   vec2 i{floor(st.x), floor(st.y)};
+   vec2 i{static_cast<float>(floor(st.x)), static_cast<float>(floor(st.y))};
    vec2 f{fract(st.x), fract(st.y)};
 
    // four corners in 2D of a tile
